+++ vue/src/index.ts
@@ -1,0 +1,11 @@
+/*!
+ * Matomo - free/libre analytics platform
+ *
+ * @link    https://matomo.org
+ * @license https://www.gnu.org/licenses/gpl-3.0.html GPL v3 or later
+ */
+
+import './BulkActions/bulkActions';
+import './CodeEditor/codeEditor';
+
+export { default as BulkActionsToolbar } from './BulkActions/BulkActionsToolbar.vue';