+++ webpack.mix.js
@@ -1,3 +1,0 @@
-let mix = require('laravel-mix');
-
-mix.js('src/js/app.js', 'javascripts/script.js');